requirements.txt
requests==2.31.0
pytest==8.1.1
pytest-xdist==3.5.0
selenium==4.20.0
allure-pytest==2.13.5
flake8==7.0.0